refactor: move SYSTEM_HANDLE_TABLE_ENTRY_INFO_EX to Windows.Win32
